Add possibility to merge pull requests with SHA
This feature allows to merge a Pull Request giving a SHA-Hash of the
Pull Request's head. If the given SHA doesn't match the actual head of
the Pull Request on GitHub, the merge fails, and BAD_HEAD is returned.
Instead of throwing an exception if the merge fails, NOT_MERGEABLE is
returned. On a successful merge, SUCCESS is returned. IOException is
thrown if any other unexpected status is returned by GitHub.